The entirety of the project so far
Classical commit to the first update, this is indeed a new major project that I've been working on for a few months or maybe longer actually. It's a sort of sandbox to test certain things that I want to do in the future in the main game, dynamic IK animation, ragdolls, body deformation and dismemberment, particle effects, vehicles and game loop shifting, environments, sounds, multiplayer, split screens, guns, questing, function passing for dynamic logic. I've tested or implemented all of these things in this project to some degree and learned a lot of things that I hope will be a boon to my main project. Especially the dynamic animation and equipment and rag dolling. basically, I want to make a very good, solid system of player (and other creatures less so) control and management that reacts fluidly and uses logic to animate and show equipment, actions, and status's like dismemberment (vs drawn animation). This is what this project is mainly for and it's honestly gone very well, some rough patches for sure but the ragdoll is interesting, the animation is pretty, the equipment is quite solid, the guns and player management is serviceable. I cannot doubt the success of this experiment. Perhaps I will turn this into something in the future, a raven field esque sandbox shooter with a focus on pushing the standard shenanegans to it's logical conclusion, huge explosion, destruction and bodily damage systems, and beautiful animation, plus vehicles and toys. That's all I really want out a game like this and what I need for testing anyway. So one day, maybe it'll turn out. Until then I look forward to improving the wheel and pushing the envelope of my design in those ways. Oh yeah! About half this is the input library... Thanks Juju. That's about it, later!