Project quickly changing goals lol
So this is almost entirely a procedural generation system, though there's some small adjustments to the desert game core. For example, enemies now shoot at other enemies in a very basic way, just for fun, i guess.. Also the helicopter particles were changed to something still ugly but less annoying... I tried man.

The big deal here though is the proc generation, now the perlin map is used by the game to also generate structures. Cool, so no more randomly placed junk right? Well it's not just color mapping, this is infinite, seeded, biome specific generation. I mean, we're approaching minecraft levels of custom generation. I can have trees (oh I added physics enabled trees that can be broken and procedurally generate by the way) and grass and bushes and all that connected vine work I described for space game happen all based on a biome map of perlin noise that get's preserved and regenerated when you come back to that spot... It's very early stages but the system works perfectly at the moment, it will still double spawn since the deleting isn't aligned with the creating but hey. The plan is to add enough foliage and other things into the generation system so that the world looks dense and interesting then I'll return to the issues of performance and interactability. I think a shader/surface pair should be excellent for making the grass and small things move and I can just straight up use force physics that I've made to move and knock around the bigger things like trees/cactus/vines/buildings ect. These two in conjunction should give me good coverage of complex physics and bulk control both being reasonably performant. Of course there's tons of optimization that I could do too. But that's the theory, and I know for a fact that it's doable, it's just going to be a lot of awkward testing and layering of different things to get everything working like I'm imagining it.. So, check back in six months I guess. This isn't that dissimilar to what I wanted to do in space game too, layered and proc generated planet environments with layers of different visuals connected smoothly with building nodes and interconnected elements. Anyway, rant over, I'm going to be busy with school for a few days so there you go.