not working, trying to get random systems working
Trying to make the ik legs with the step timing system more sophisticated while not making them a nightmare to use, I just want the legs to work and take the steps when they're supposed to but that's really hard it turns out. But anyway. I also tried a few times to get the particle system wrapper working because it's... close? I think? It's the classic finicky math thing where there's a bunch of interconnected values that all need to line up but the main problem is that there's no way to get the depth of a particle system ):< So it's been very hard to debug.

Various other experimental things in progress i guess