It just won't work But also it is mostly working
It does work, there's some issues I know of, one, it's not updating the systems when you do the update function (the whole point??) and also the macro for border ratio cannot be edited or the depthing will break and I really tried to get the depth to work but I think it's broken somewhere else, my guess is the setting is off somewhere and fixing it in the depth is only a bandaid sort of thing. Also changing depth origin breaks it. Cool... But it's 90% there I suppose.