Progress, progress, weapon maw poss, 3d additions
So this makes weapons and limb drawing use 3d positions which is a bit annoying to set up but should almost certainly simplify the usage of the system moving forward, turns out doing this properly instead of trying to jerry rig it to look correct but not actually be is hard, who would have thought. This also adds a couple polish things and a position for the tip of the gun, based on variable offsets. So the bullets now come out of the right part of the gun and whatnot. Looking pretty solid ngl, still very rough and needs several things stripped and remade (legs, LEGS, LEGS!!!) but it's getting there I guess. Having the different parts of the system be functions is really nice for organization and not having to worry about duplicating code is a god send, like i said, it's a disaster not to so I'm feeling the benefits now that it's changed. Hopefully more to come soon, I feel that it's on the precipise of a metamorphosis but garbage to greatness, things are falling into place, it's looking proper, I'm rambling and repeating myself because I've spent 9 MONTHS trying to make god damn stepping legs and held weapons in this stupid game... But, we prevail, in the end we prevail.