Structs working for sure now, couple small tweaks
Fixed some variable finding and sorted out the image/sprite getting for rotations. I also lag tested a few different set ups and it seems like this new struct system is totally fine on performance, I didn't compare to the array set up but with a couple thousand arms (limbs) all drawing it was still getting around 50-100 in yyc so.. should be fine, there's lots of ways to speed that up if need be. But I doubt I'll have 2000 limbs all sorted into a creature any time soon so. We chillin.