Confused but also good progress on body control
Head now tilts towards aimed direction (more generally there's ways to set the angle and other values of visual components to code values (like aiming direction). This can be used for example to make a head visual align with view cones in main game or with weapons in this game. I also adjusted a bunch of small values for setting up spines and ATTEMPTED to make the spines work, eg the weapons, components, limbs ect use the spine position instead of the instance position when positioning in the game. This has not been super succesful and something about the weapon position is still broken, and not respecting height somehow, or it's being set twice with height but in some other way because all the little adjustments I tried didn't work properly so idk.