twas height offsets all along, working now... ish
I've put height offsets into the component drawing more cleanly and made a couple more things use height (the components for spines, that is, and spines themselves for that matter), and grouped it all together for drawing. The weapon tips are still offset to hell, probably some mismatch between the spineX and the gun x somehow.., and the legs are now showing their age with not being a component part and not respecting any offsets from there. So fix the gun and redo the legs as components and also make the step positions more properly fixed to points and whatnot like I did with the dinosaurs but better.. That's the next step before it's almost done for the big stuff and needs tweaking