new style of legs set up as component limbs, WIP
It's no where close to working but I've got the legs that use the fixed step rate and ground position choosing built into the player object as components, so there's an array for legs and an array for all other limbs at the moment. The best I got is the legs dragging behind because the stepping is broken but the stepping was already broken so... And there's going to have to be a balance between manual control of variables and procedural implementation of things, like making physics nodes should be fairly simple in terms of defining the type and shape of the nodes or extra visual components, these are things that shouldn't need much oversight whereas legs and arms and the head for example have values within their components that are vital to be able to change from else where in the code to make work how you'd expect (stepping, holding things, looking with a look direction value..) So I give some of these components names and I think it works well, again, time will tell where the right balance is and if this system works or will have to be adjusted. Plus lots of additions, as always!