tree node knockback and some minor leg fixes
Nukes now knock back trees which is cool but the generation doesn't zoom out for it so it's a bit of a mess, also, the legs for the player are now "working", but don't look right at all and have some weird positioning and timing issues that I've bandages over for now but i'm not really sure what the correct method here is. It's also quite annoying to debug so i've barely figured anything out.