Trying to debug is not going well
The problem is that the legs are being distance clamped because they're using height now, ergo with the extra dimension the old distance is now less visible and so the legs more quickly get clipped off but because of the way that the clamping works they get pulled in over and over but the foot placement keeps setting them instantly back so instead of just getting restricted they have this weird ghost approaching of the point they think they want and just kind of float towards the point. Something about the positioning is broke.