Making some progress finally (my own errors)
Turns out some of the biggest problems were from me trying to handle the distance clamping, I had a line in there that clamped the distance of the legs to the limb len, aka instead of going out a thigh plus a shin it was getting pulled back to just one segment or half of what it should have been. Because I also have 1 or two other places where it has to clamp the distance this was pretty confusing but I've got it sorted now.. I also got annoyed and imported the actual 3d rendering script from main game so the leg segments are now actual, accurate 3d, however because the camera is (in theory) 45 degrees this makes basically no difference since i was just sorta baking the 3d values in before (at .7 vs .71 or something). Anyway, fixing the clamping plus the rendering plus a couple tiny adjustments trying to see if it's a model issue and not a render issue it looks a decent amount better now. It's still not like, correct looking, and I'm not really sure why / what's wrong but oh well. I also added a couple spine variables in there for completeness sake. It may be a bit of a mess from a performance perspective but whatever man, I'll deal with optimization when it's done.