sink when changing speed/dir, cleanup, recentering
The character now sinks down when making a sharp turn or changing speeds quickly / at all (more for more change), I also tuned the values for step placement and cadence and added a way for them to self cull their stance when not moving. So now they step more realistically when getting around and don't have the issue of step clumping anymore. They also replace their steps to a more neutral position when standing still, for the most part, they sometimes end up with a slight lean but I probably need to adjust the values.. But still, no more stanky legging when standing. They also have nice height when standing, walking and running, standing all the way up at a stand, slightly lower to walk and hunched when running. All of this is pretty cool but really quite simple code wise, a little ribbon of detail makes the whole illusion of the body look different so, eh. It's hard to work on, still. I also cleaned up some nonsense (for lag reasons on my laptop) and made some better debug overlays, though still simple. The next step should definitely be fixing the 3d display of the legs, they look completely busted when turned or lower than 90 degrees. They pop in and out of each other and all kinds of nonsense. But it's looking better, hopefully sustainable...