RAAAAAAHHHHHHH
Sorry, but apparently the forward portion of an angle relative to a point is just the dot product, who knew? Well, lots of people I'm sure but not me. But... it means that the solution was easy and now the legs finally, finally... work. (render correctly / place knees correctly..) Now to adjust how they move and get around and all the mechanics of the thing.. But finally the core is just simply done. No more rendering pipeline issuesss