Huge changes to code hierarchy ECS'ing body stuff!
This isn't working particularily well but basically I'm classing the body drawing stuff (now that it's mostly working) to be taken up with functions and such instead of hard coded to the player. The next bit of work will probably involve a lot of instancing variables that were local so that they can be moved between functions and generalized to different objects. (eg aimDir instead of dirToMouse) because enemies can aim and so can the player but they won't be aiming at the mouse..) But it's good, decently working and all building up to having the full body system be working for enemies just like players. This is one the 3 or 4 steps I described way back for this game being finished, inheritance of body models as a component that can be applied to anything you like instead of hard code. The other cool thing I'll do soon hopefully is bringing some of the enemy abilities (ragdolls and limb removal) over to this system and unify all the different types of body functionalities with toggled options or thresholds like limbDismembermentThreshold or canBeRagdolled ect instead of it only being possible for enemies / the old body system... Gotta get everything gathered and cleaned!!