Fixed visual sorting and invisible bug!
So yeah there were two very annoying issues that both had to be picked away at to figure out, issue one is that the array sorting function in game maker seems to round to whole numbers?? So I just multiplied the sort difference it uses by 100000 or something to make it not decimal, I also made it so ties are solved by sorting by index, not random placement (which is the default) so now they don't pop forward and backward and the window of equality is really small (not 1 round wtf). The other issue is much more underwhelming, for the loop that draws the in behind body parts, breaks, then draws the in front parts, the loop wasn't breaking because it was in a with() call so it was just not looping correctly. Very lame error to spend a few hours half messing with but whatever. It is fixed now...