Made the body debug colors and fixed lost fix from before
What does fixed lost fix blah blah mean? The fix I did back in time that fixed  the sorting and popping issue was accidentally removed when I switched the sort script to the mdl_scripts stuff and so I had to repaste the fix into the new script. No idea why I messed that up but it's fixed now.

For reference, the two big issues THAT NEED TO BE FIXED FOR THE PROJECT TO BE SECURE, are the step timing being asymptoted to large numbers when changing speeds near 0, and the logic for placing and starting steps, not only is starting and stopping steps not understandable but also is very problematic, being very hard to change or fix. So I need to remake the logic for step management, especially if I'm hoping to have support for non 2 legged creatures at some point... DO THOSE NEXT TIME